Moves BySource into "elements" folder
