+++ web/src/features/charts/bar-breakdown/elements/BySource.tsx
